Rename homepage component
