+++ packages/website/src/layout/App/index.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { useDispatch } from "react-redux";
 
 import { NotFound } from "../../routes/NotFound";
-import HomePage from "../../routes/Homepage";
+import HomeMap from "../../routes/HomeMap";
 import ReefRoutes from "../../routes/ReefRoutes";
 import About from "../../routes/About";
 import RegisterSite from "../../routes/RegisterSite";
@@ -44,7 +44,7 @@ function App() {
         <div id="app">
           {render && (
             <Switch>
-              <Route exact path="/" component={HomePage} />
+              <Route exact path="/map" component={HomeMap} />
               <Route exact path="/about" component={About} />
               <Route exact path="/register" component={RegisterSite} />
               <Route exact path="/apply" component={Apply} />
